requirements.txt
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
Django==2.2.1
django-cors-headers==3.0.1
django-markdownx==2.0.28
django-templated-mail==1.1.1
djangorestframework==3.9.4
djoser==1.5.1
httpie==1.0.2
idna==2.8
Markdown==3.1.1
Pillow==6.0.0
Pygments==2.4.0
pytz==2019.1
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.3
